requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
sqlalchemy==2.0.*
psycopg[binary]==3.2.*
passlib[bcrypt]==1.7.*
PyJWT==2.8.*
python-multipart==0.0.*
pydantic==2.*
python-dotenv==1.0.*